+++ packages/arcgis/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @openfn/language-arcgis
 
+## 1.0.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @openfn/language-common@3.1.2
+
 ## 1.0.0
 
 ### Major Changes