+++ packages/twilio/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @openfn/language-twilio
 
+## 1.0.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @openfn/language-common@3.1.2
+
 ## 1.0.4
 
 ### Patch Changes